+++ backend/src/main/java/unpsjb/labprog/backend/business/repository/EsquemaTurnoRepository.java
@@ -6,8 +6,8 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import unpsjb.labprog.backend.model.EsquemaTurno;
 
-public interface EsquemaTurnoRepository extends JpaRepository<EsquemaTurno, Long> {
+public interface EsquemaTurnoRepository extends JpaRepository<EsquemaTurno, Integer> {
 
-    List<EsquemaTurno> findByStaffMedicoId(Long staffMedicoId);
+    List<EsquemaTurno> findByStaffMedicoId(Integer staffMedicoId);
 
 }
