+++ backend/src/main/java/unpsjb/labprog/backend/business/repository/MedicoRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import unpsjb.labprog.backend.model.Medico;
 
-public interface MedicoRepository extends JpaRepository<Medico, Long> { 
+public interface MedicoRepository extends JpaRepository<Medico, Integer> { 
     boolean existsByDni(Long dni);
 
     boolean existsByMatricula(String matricula);
